requirements.txt
flask==3.0.2
flask-cors==4.0.0
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.2.1
google-generativeai==0.3.2
python-dotenv==1.0.1 